Refactor model package structure in tests
