+++ src/test/kotlin/dev/stalla/model/googleplay/ExplicitTypeTest.kt
@@ -1,4 +1,4 @@
-package dev.stalla.model
+package dev.stalla.model.googleplay
 
 import assertk.assertThat
 import assertk.assertions.contains
@@ -7,7 +7,6 @@ import assertk.assertions.isNotNull
 import assertk.assertions.isNull
 import assertk.assertions.prop
 import dev.stalla.arguments
-import dev.stalla.model.googleplay.ExplicitType
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ArgumentsProvider
