+++ src/test/kotlin/dev/stalla/model/itunes/ItunesCategoryTest.kt
@@ -1,4 +1,4 @@
-package dev.stalla.model
+package dev.stalla.model.itunes
 
 import assertk.all
 import assertk.assertThat
@@ -10,7 +10,6 @@ import assertk.assertions.isNotNull
 import assertk.assertions.isNull
 import assertk.assertions.prop
 import dev.stalla.arguments
-import dev.stalla.model.itunes.ItunesCategory
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ArgumentsProvider
